Agregadas las clases Item y Nivel, con algunos metodos, pero no están terminadas. Cambios en arma y en plano.

+++ trunk/2042-full/juego/src/juego/Arma.java
@@ -3,7 +3,7 @@ import excepciones.*;
 
 public abstract class Arma extends ObjetoUbicable {
 	
-	float danio;
+	int danio;
 	boolean usada;
 	boolean origenAlgo42;
